Exercise 6 : BOKEH PLOTS
Description: This program will create four plots from data in a matrix (list of lists or pandas). 
Input: is the file SSE_Faculty.csv, containing the courses taught by the faculty of the School of Systems 
and Enterprises from Spring '19 to Spring '23 (modified version of the actual data). 
Per each Academic Year, there is "Load" (number of classes taught by the faculty); "Target" (number of 
classes the faculty was assigned to teach); "Balance" (Load - Target).