Final Python Project
My final project consists of the data analysis of the propagation of the COVID-19 virus in the USA. Along with this I also point out all the major factors that lead to its rapid spread. My analysis is completely using pandas to generate Matplot graphs as Ill as Bokeh plots for data visualization.